Save function moved to save repository

+++ src/Managers/SaveManager.php
@@ -3,16 +3,19 @@
 namespace Managers;
 
 use Repository\Interfaces\SaveRepositoryInterface;
+use Repository\Interfaces\UserRepositoryInterface;
 use DTOs\ShowSavesDto;
 use Validators\Result;
 
 class SaveManager
 {
     private $saveRepository;
+    private $userRepository;
 
-    function __construct(SaveRepositoryInterface $saveRepository)
+    function __construct(SaveRepositoryInterface $saveRepository, UserRepositoryInterface $userRepository)
     {
         $this->saveRepository = $saveRepository;
+        $this->userRepository = $userRepository;
     }
 
     /**
@@ -26,8 +29,23 @@ class SaveManager
         return $this->saveRepository->getSaveIdsWithNames($userId);
     }
 
+    public function createSave(int $userId, string $saveName, $boardState, $elapsedTime) : Result
+    {
+        if (!$this->userRepository->userIdExists($userId))
+        {
+            return new Result(false, ["User id ${userId} not exists"]);
+        }
+
+        return $this->saveRepository->createSave($userId, $saveName, $boardState, $elapsedTime);
+    }
+
     public function deleteSave(int $userId, int $saveId): Result
     {
+        if (!$this->userRepository->userIdExists($userId))
+        {
+            return new Result(false, ["User id ${userId} not exists"]);
+        }
+
         if(intval($this->saveRepository->getSaveData($saveId)["user_id"]) !== $userId)
         {
             return new Result(false, ["Save not found"]);
@@ -38,6 +56,11 @@ class SaveManager
 
     public function deleteUserSaves(int $userId): Result
     {
+        if (!$this->userRepository->userIdExists($userId))
+        {
+            return new Result(false, ["User id ${userId} not exists"]);
+        }
+
         return $this->saveRepository->deleteUserSaves($userId);
     }
 }
